back after lil vacation>
added new spawn points for some jumpy enemies, also changed the height of all the spawn points and waypoints to 2. since the enemies are on Y:2.
added few new jumpy enemies. one in the SQLeft side at the second level with 8 spawn points and one in the main hall with 3 spawn points.

changed a few things in the active behaviour script in the enemies. now they have their own sets of spawn points and number. for a dynamic handling of spawning the jumpy ones again.

on the verge of decoupling lots in the journal progression for more oversight. right now the enemies that are enabled after an object is picked up is manually in the editor done though i want to change it a bit more dynamic for easier adding of enemies. might not do it at all since there aren't that many enemies at all.

NOTE: gotta playtest again.
gotta make props so the player can hide behind them for a moment. especialy in the mess and audit.
gotta disable the creep and sound of some enemies when the player is in a certain area so they wont be noticed behind a thick wall hmmm
gotta make the jumpy lights, light up more when activated for a brief 'shock' and sounds for them.